@@ -3,33 +3,32 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^24.0.0",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.49",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-select": "^3.0.21",
-    "@types/styled-components": "^5.1.3",
-    "i18next": "^19.7.0",
+    "@testing-library/jest-dom": "^6.0.1",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.5.3",
+    "@types/node": "^20.5.1",
+    "@types/react": "^18.2.20",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-select": "^5.0.0",
+    "@types/styled-components": "^5.1.26",
+    "i18next": "^23.4.4",
     "i18next-xhr-backend": "^3.2.2",
-    "mobx": "^6.0.0",
-    "mobx-react-lite": "^3.0.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-i18next": "^11.7.3",
-    "react-scripts": "3.4.3",
-    "react-select": "^3.1.0",
-    "styled-components": "^5.2.0",
-    "typescript": "~3.7.2"
+    "mobx": "^6.10.0",
+    "mobx-react-lite": "^4.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^13.1.2",
+    "react-select": "^5.7.4",
+    "styled-components": "^6.0.7",
+    "typescript": "~5.1.6"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "e2e":"cypress open"
+    "start": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "cy:open-e2e": "cypress open --e2e --browser chrome",
+    "e2e": "start-server-and-test start http-get://localhost:3000 cy:open-e2e"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -47,6 +46,9 @@
     ]
   },
   "devDependencies": {
-    "cypress": "^5.3.0"
+    "@vitejs/plugin-react": "^4.0.4",
+    "cypress": "^12.17.4",
+    "start-server-and-test": "^2.0.0",
+    "vite": "^4.4.9"
   }
 }
