@@ -1,5 +1,4 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 import { ThemeProvider } from "styled-components";
@@ -7,19 +6,21 @@ import { theme } from "./styles/theme";
 import GlobalStyle from "./styles/GlobalStyle";
 import { UserStore } from "./store/UserStore";
 import { RootStateProvider } from "./store/RootState";
+import { createRoot } from 'react-dom/client';
 import "./i18n";
 
 const userStore = new UserStore();
 export const RootStoreContext = React.createContext(userStore);
-
-ReactDOM.render(
+const domNode = document.getElementById('root') 
+if (!domNode) throw new Error('Failed to find the root element');
+const root = createRoot(domNode)
+root.render(
   <RootStateProvider>
     <ThemeProvider theme={theme}>
       <GlobalStyle />
       <App />
     </ThemeProvider>
-  </RootStateProvider>,
-  document.getElementById("root")
+  </RootStateProvider>
 );
 
 serviceWorker.unregister();
